+++ CHANGELOG.md
@@ -7,3 +7,4 @@
 * `currentPlayer` is now a function, not a property.
 * `territoryScore()` and `areaScore()` are replaced with `score()`. A new `setup()` option allows defining the scoring in use for a game.
 * Superko support with a configurable `koRule` option.
+* Support for equivalence scoring (with pass stones) as a scoring option.